Bugfix for Spryker secrets not appearing in generated env files.
This commit fixes a bug in the generator, which caused the generated env files to not contain the Spryker secrets,
e.g. SPRYKER_OAUTH_KEY_PRIVATE. The bug occured when the used deploy.yml file contained a top-level 'project:' yaml
node. In for example zed.env.twig generator template, the common.env.twig was being included with the `project`
variable being extracted, leading to all child-nodes of `project` then being top-level nodes inside of common.env.twig.
This `project` variable is the `$projectData` from generator/index.php, and hence contains a representation of the yaml
structure of the deploy file.

In other words: If we extract `project` when including the common.env.twig, then all top-level yaml nodes from the deploy
file will be standalone twig variables in the common.env.twig. This means, that if we have a 'project:' yaml node
in the deploy file, then this yaml node will override the original `project` variable itself. This should not be
expected behaviour, and the author of this PR could not find any valid reason why this should be done.

So this commit removes the extraction of the `project` variable and fixes the only one (1) occurence in the common.env.twig
file where an extracted variable actually was being used. This variable is now being directly addressed as `project['variable']`.