Fix docker-compose version check bug.
With newer Version docker the "docker-compose version --short" command gives out a version number string with "v" at the beginning, so the later comparison failed with this misleading error message:  
`Docker Compose version v2.1.1 is not supported. Please update Docker Compose to at least 1.22.0. `
With the help of "tr" to remove the leading "v" will fix this problem.